require af-scikit-learn; install it first! TODO fix setup.py
setup.py uses the name af-scikit-learn to identify the
project. the package must therefore be installed before
this package.

as a TODO, we must target af-scikit-learn with its github
url while not using both packages in their editable form.
referencing only by package name enables af-scikit-learn
to work in editable mdoe. once changed to a url, it cannot
be in editatle mode